requirements.txt
# Required Libraries
#BeautifulSoup
newspaper3k
lxml_html_clean
googlesearch-python
duckduckgo-search
tensorflow
nltk
scikit-learn
googletrans
gradio
groq
spacy
python-dotenv
textblob
gtts
cohere
protobuf==4.25.3
transformers==4.38.2
